+++ common/src/python/outputs/errors.py
@@ -71,6 +71,12 @@ preprocess_errors = {
     SysErrorCodes.MISSING_UDS_I4:
     "Participant must have an existing Initial UDSv4 Visit Packet (PACKET=I4) "
     + "submitted before the Follow-Up Visit Packet (PACKET=F)",
+    SysErrorCodes.DUPLICATE_VISIT:
+    "Duplicate record with the same visit date exists in the batch CSV file " +
+    "for this participant",
+    SysErrorCodes.LOWER_VISITNUM:
+    "Packet with higher visit date (VISITDATE) " +
+    "must also have a higher visit number (VISITNUM)",
 }
 
 
@@ -491,10 +497,15 @@ def update_error_log_and_qc_metadata(*,
                   f'{destination_prj.group}/{destination_prj.label}: {error}')
         return False
 
+    # if error data already exists, append to data
+    existing_errors = info.get('qc', {}).get(gear_name, {}) \
+        .get('validation', {}).get('data', [])
+    existing_errors.extend(errors)
+
     info["qc"][gear_name] = {
         "validation": {
             "state": state.upper(),
-            "data": errors
+            "data": existing_errors
         }
     }
     try: