+++ docs/csv_subject_splitter/CHANGELOG.md
@@ -2,13 +2,21 @@
 
 All notable changes to this gear are documented in this file.
 
-## TBD
+## 1.0.3
 
-- n/a
+* Changes splitter so that file is uploaded for each row rather than saving all rows and then uploading
 
-## 1.0.1 (Unreleased)
+## 1.0.2
 
+* Updates `ProjectAdaptor.add_subject` to return the subject if it already exists
+* Updates `uploader.JSONUploader` to remove the `allow_updates` parameter - will always update instead
+
+## 1.0.1
+
+* Updates `hierarchy_labels` config to be type `string` instead of type `object`
+* Updates `uploads.uploader.JSONUploader` to accept a `allow_updates` parameter (defaults `false`) and, if `false`, throw an error if the subject already exists
 * Refactors to use `InputFileWrapper.get_parent_project`
+* Removes `assert parameter_store` statement
 
 ## 1.0.0